requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.6.4
prometheus-client==0.20.0
pdfminer.six==20231228
numpy==1.26.2
scikit-learn==1.3.2
rapidfuzz==3.5.2
python-multipart==0.0.9
faiss-cpu==1.8
joblib==1.3.2
packaging==23.2
groq==0.9.0
requests==2.32.3